+++ src/components/Renderer/ui/index.js
@@ -1,3 +1,6 @@
 
+import TransactionLink from "./Link/Transaction";
 
-export {}
+export {
+  TransactionLink,
+}