+++ src/index.js
@@ -21,6 +21,11 @@ ReactDOM.render(<PrismaCms
     apiQuery: `{
       user:me{
         ...UserNoNesting
+        EthAccounts {
+          id
+          address
+          balance(convert:ether)
+        }
       } 
     }
     ${UserNoNestingFragment}